backend/requirements.txt
autopep8==2.0.0
beautifulsoup4==4.11.1
bunch==1.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
Flask==2.2.2
Flask-Cors==3.0.10
google==3.0.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.21.9
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
simple-colors==0.1.5
six==1.16.0
slack==0.0.2
slack-sdk==3.19.4
soupsieve==2.3.2.post1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
